requirements.txt
Flask==1.0.2
requests==2.20.0
mysql-connector==2.2.9
